+++ hookee/conf.py
@@ -12,6 +12,7 @@ template = {
     "default_route": confuse.String(default="/webhook$"),
     "default_route_methods": confuse.String(default="^(GET|HEAD|POST|PUT|DELETE|PATCH|OPTIONS|TRACE|CONNECT)$"),
     "port": int,
+    "no_tunnel": confuse.OneOf([True, False], default=False),
     "subdomain": confuse.String(default=None),
     "region": confuse.Choice(["us", "eu", "ap", "au", "sa", "jp", "in", "us-cal-1"], default=None),
     "domain": confuse.String(default=None),